Add config-server-encrypt-value
* Generalise the description of --skip-ssl-validation

* Record an inconsistency in the contents of access token values in the code and in
  https://www.pivotaltracker.com/story/show/143022981

Note that fake_cli_connection.go cannot be regenerated easily. Regenerating it
as part of this project fails because of
https://github.com/maxbrunsfeld/counterfeiter/issues/50. Meanwhile the CLI
project has moved on and we would need to upgrade to pick up regenerated fakes
in the CLI project. So, for the moment, we avoid using new counterfeiter
features to work around these problems.

[#138076761]