service-registry-enable/disable
service-registry-enable issues a PUT to apps/appID/instanceID/status?value=UP
following the note about fallback status in [1]. This works around the re-
registration bug #136954687. (If the PUT is issued without the fallback status
(i.e. to apps/appID/instanceID/status), the service registation status remains
at UNKNOWN and the applications fails to re-register.)

Note: the alias of the service-registry-deregister command is changed from
"srd" to "srdr" to avoid confusion with service-register-disable.

[1] https://github.com/Netflix/eureka/wiki/Eureka-REST-operations

[#133142105, #136954687]