Progress reporting
The following commands now report progress as they proceed:
* service-registry-enable
* service-registry-disable
* service-registry-deregister

The other commands do not produce output until they are just about to complete
successfully.

Note: an earlier version of this commit passed all tests when using `govendor
test +local` or when running under IntelliJ and failed a test because of
missing colour codes in the expected string under `ginkgo -r` or when
debugging under IntelliJ. This commit does not attempt to straighten out this
kink, but all the tests now pass in all environments.

[#137440279]