Add a Git repository to Config Server via CLI
* Adds `config-server-add-git` CLI command
* Refactor repo manipulation command tests to share examples

I think over time we'll adapt this to support the rest of the
configuration parameters that can be supplied when specifying a git repo
to add. But for now, the story asks for support of a command of the form

```
cf config-server-add-git config-server URI
```

so this is what I've gone for. It might be that we want to support this
simple URI style to allow quick adds, and a more verbose JSON blob for
more involved configurations.

connected to pivotal-cf/spring-cloud-service-broker#320